+++ applications/backend/views/_partial/footer.php
@@ -1,0 +1,7 @@
+
+<footer class="main-footer">
+	<div class="pull-right hidden-xs">
+		<b>Version</b> <?php echo WEBSITE_VERSION; ?>
+	</div>
+	<strong>Copyright © 2015 <a href="<?php echo base_url(); ?>"><?php echo COMPANY_NAME; ?></a>.</strong> All rights reserved.
+</footer>