+++ applications/backend/views/_partial/header.php
@@ -1,26 +1,21 @@
 
-<header class="header">
-
-	<a href="<?php echo site_url(); ?>" class="logo">
-		Backend System
-	</a>
-
+<header class="main-header">
+	<a href="<?php echo site_url(); ?>" class="logo"><b>Backend System</b></a>
 	<nav class="navbar navbar-static-top" role="navigation">
-		<a href="#" class="navbar-btn sidebar-toggle" data-toggle="offcanvas" role="button">
+		<a href="#" class="sidebar-toggle" data-toggle="offcanvas" role="button">
 			<span class="sr-only">Toggle navigation</span>
 			<span class="icon-bar"></span>
 			<span class="icon-bar"></span>
 			<span class="icon-bar"></span>
 		</a>
-		<div class="navbar-right">
+		<div class="navbar-custom-menu">
 			<ul class="nav navbar-nav">
 				<li class="dropdown user user-menu">
 					<a href="#" class="dropdown-toggle" data-toggle="dropdown">
-						<i class="fa fa-user"></i>
-						<span><?php echo $user['full_name']; ?> <i class="caret"></i></span>
+						<span class="hidden-xs"><?php echo $user['full_name']; ?></span>
 					</a>
 					<ul class="dropdown-menu">
-						<li class="user-header bg-light-blue">
+						<li class="user-header">
 							<p><?php echo $user['full_name']; ?></p>
 						</li>
 						<li class="user-footer">
@@ -36,5 +31,4 @@
 			</ul>
 		</div>
 	</nav>
-
 </header>