+++ applications/backend/views/_partial/menu.php
@@ -1,5 +1,8 @@
 
 <ul class="sidebar-menu">
+
+	<?php // Section 1 ?>
+	<li class="header">MAIN NAVIGATION</li>
 	<?php
 		foreach ($menu as $parent => $parent_params)
 		{
@@ -10,7 +13,7 @@
 			if ( empty($parent_params['children']) )
 			{
 				$active = ( current_url()==$url ) ? 'active' : '';
-				echo "<li class='$active'><a href='$url'><i class='$icon'></i> <span>$name</span></a></li>";
+				echo "<li class='$active'><a href='$url'><i class='$icon'></i> $name</a></li>";
 			}
 			else
 			{
@@ -23,7 +26,7 @@
 				$list_arrow = ($parent_active) ? 'fa-angle-down' : 'fa-angle-left';
 				echo "<li class='treeview $parent_active'>
 					<a href='#'>
-						<i class='$parent_icon'></i> <span>$parent_name</span><i class='fa pull-right $list_arrow'></i>
+						<i class='$parent_icon'></i> $parent_name <i class='fa $list_arrow pull-right'></i>
 					</a>";
 
 				// child items
@@ -32,11 +35,16 @@
 				foreach ($parent_params['children'] as $name => $url)
 				{
 					$child_active = (current_url()==$url) ? "class='active'" : '';
-					echo "<li $child_active><a href='$url' style='margin-left: 10px;'><i class='fa fa-angle-double-right'></i> $name</a></li>";
+					echo "<li $child_active><a href='$url'><i class='fa fa-circle-o'></i> $name</a></li>";
 				}
 				echo "</ul>";
 				echo "</li>";
 			}
 		}
 	?>
+
+	<?php // Section 2 ?>
+	<li class="header">USEFUL LINKS</li>
+	<li><a href="<?php echo base_url(); ?>" target='_blank'><i class="fa fa-circle-o text-info"></i> Frontend Site</a></li>
+
 </ul>