+++ applications/frontend/views/_partial/foot.php
@@ -1,7 +1,7 @@
 
 	<div class="footer">
 		<div class="container">
-			<p class="text-muted">Copyright &copy; 2014.</p>
+			<p class="text-muted">Copyright &copy; 2015.</p>
 		</div>
 	</div>